Move VLQ out of strings
