+++ src/vlq.ts
@@ -1,5 +1,8 @@
 import type { StringReader, StringWriter } from './strings';
 
+export const comma = ','.charCodeAt(0);
+export const semicolon = ';'.charCodeAt(0);
+
 const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/';
 const intToChar = new Uint8Array(64); // 64 possible chars.
 const charToInt = new Uint8Array(128); // z is 122 in ASCII
@@ -44,4 +47,9 @@ export function encodeInteger(builder: StringWriter, num: number, relative: numb
   } while (delta > 0);
 
   return num;
+}
+
+export function hasMoreVlq(reader: StringReader, max: number) {
+  if (reader.pos >= max) return false;
+  return reader.peek() !== comma;
 }
